+++ main.rb
@@ -1,51 +1,15 @@
 require_relative './app'
-
-def menu
-  puts 'Welcome to School Library app'
-  puts 'Select an option by entering a number: '
-  puts '1) List all books'
-  puts '2) List all people'
-  puts '3) Create a person'
-  puts '4) Create a book'
-  puts '5) Create a rental'
-  puts '6) List all rentals for a given person id'
-  puts '7) Exit'
-end
+require_relative './menu'
+require_relative './menu_executer'
 
 def main
-  print 'welcome!'
-  app = App.new
+  menu = Menu.new()
+  app = App.new()
+  executer = MenuExecuter.new()
   loop do
-    puts menu
+    puts menu.menu_list
     option = gets.chomp.to_i
-    menu_executer(option, app)
-  end
-end
-
-def options(option, app)
-  case option
-  when 1
-    app.list_of_all_books
-  when 2
-    app.list_of_all_people
-  when 3
-    create_person(app)
-  when 4
-    puts app.create_book
-  when 5
-    app.create_rental
-  when 6
-    app.list_of_all_rentals
-  end
-end
-
-def menu_executer(option, app)
-  if option == 7
-    puts 'Thank you for using this app!'
-    exit
-  else
-    options(option, app)
-
+    executer.menu_executer(option, app)
   end
 end
